Overhauled the ProgressObserver and its friends
The previous ProgressObserver infrastructure was unable to deal with
two or more simultaneous operations. Moreoever it solely relied on
Glib::Dispatcher which required the observed operation to be done in a
separate thread.

A ProgressDialog has been to the Application, and whenever a
ProgressObserver is created it is associated with this dialog. The
observer is a Gio::Cancellable, and a template that expects either a
Glib::Dispatcher or a sigc::slot as the parameter. When the operation
begins, the observer attaches a ProgressWidget to the ProgressDialog
and detaches it when cancelled or destructed. The ProgressDialog and
ProgressWidget were translated to C++ from C code written for
Nautilus.

A IProgressObserver was added as an interface to the actual templates
for use by user code:
    + IProgressObsever::is_finished
    + IProgressObsever::progress
    + IProgressObsever::set_description
    + IProgressObsever::set_fraction
    + IProgressObsever::set_total

Construction and destruction of a IProgressObserver, and
IProgressObserver::set_description are only meant to work from the
main thread they involve interactions with the GUI.

IPhotoDestination::export_photos has been renamed to
IPhotoDestination::export_photos_async and it is run in the main
thread. It is unknown whether the Brasero libraries are thread-safe or
not, and there was really no pressing need for using a separate
thread.

The IPhotoDestination::export_photo has been removed as there was no
real need for it and it was becoming too different for the Brasero and
directory destinations.

EditablePhoto::pending_ now contains triplets instead of pairs -- a
ProgressObserverPtr being the new addition. If an ongoing operation is
cancelled (or the initial Gegl::Buffer could not be created) the
entire queue is cleared and a new GdkPixbuf is not created.

A Finally was added to mimic Java's finally clause. This is useful for
invoking of g_object_unref when moving out of the local scope, or
emitting Glib::Dispatchers to indicate the end of a thread worker.

Regressions:

Removal of the temporary directory used for creating an archive when
exporting to a directory is now done synchronously in the main thread.