Basic ArchiveMaker for creating archives
Spawning file-roller for creating archives is sub-optimal. The photos
need to be copied to a directory and then the directory has to be
archived. Instead we can directly put the photos in the archive and
avoid copying them.

The ArchiveMaker is stateless and provides an asynchronous way of
creating an archive from a list of photos. Currently it only creates
ZIP archives, but other formats can be easily added.

A new dependency has been introduced on libarchive.