Added a PhotoFactory
Currently photos are instanciated in the database when a query returns
a bunch of URIs and they get destructed when the last reference is
dropped. It is possible that the same URI is represented by two
separate instances of Photo -- one in the export queue and another
created from the results of a new query. For consistency it is better
to have only one Photo instance represent a particular URI at any
time.

Instances of Photo should be created using PhotoFactory::create_photo.
The PhotoFactory is not thread-safe as we are constructing or
destructing photos only from the main thread. The PhotoFactory keeps a
std::map of uris and WeakPhotoPtrs. When the last PhotoPtr referring
to a particular Photo instance is dropped the corresponding
WeakPhotoPtr is removed from the std::map. Therefore all PhotoPtrs
should be destructed before the PhotoFactory.

Added a Thumbnailer::shutdown, which should be invoked before the
PhotoFactory is destructed.